+++ plugins/McpAutomationBridge/Source/McpAutomationBridge/Private/Domains/AssetWorkflow/Structs/McpAutomationBridge_AssetWorkflowStructsHelpers.cpp
@@ -276,4 +276,5 @@ void ForEachReferencingBlueprint(UUserDefinedStruct* S, TFunction<void(UBlueprin
     });
 }
 
+
 #endif // WITH_EDITOR